requirements.txt
fastapi
uvicorn[standard]
modelcontextprotocol
openai
python-dotenv
